Fix integer comparision compiler warnings

+++ blockwise_sa.h
@@ -988,7 +988,7 @@ template<typename TStr>
 void KarkkainenBlockwiseSA<TStr>::nextBlock(int cur_block, int tid) {
 #ifndef NDEBUG
 	if(this->_nthreads > 1) {
-		assert_lt(tid, this->_itrBuckets.size());
+		assert_lt((size_t)tid, this->_itrBuckets.size());
 	}
 #endif
 	std::vector<TIndexOffU>& bucket = (this->_nthreads > 1 ? this->_itrBuckets[tid] : this->_itrBucket);
@@ -998,7 +998,7 @@ void KarkkainenBlockwiseSA<TStr>::nextBlock(int cur_block, int tid) {
 	}
 	assert(_built);
 	assert_gt(_dcV, 3);
-	assert_leq(cur_block, _sampleSuffs.size());
+	assert_leq((size_t)cur_block, _sampleSuffs.size());
 	const TStr& t = this->text();
 	TIndexOffU len = (TIndexOffU)t.length();
 	// Set up the bucket
@@ -1069,7 +1069,7 @@ void KarkkainenBlockwiseSA<TStr>::nextBlock(int cur_block, int tid) {
 			}
 			if(!last) {
 				// Not the last bucket
-				assert_lt(cur_block, _sampleSuffs.size());
+				assert_lt((size_t)cur_block, _sampleSuffs.size());
 				hi = _sampleSuffs[cur_block];
 				zHi.resize(_dcV);
 				zHi.assign(zHi.size(), 0);
@@ -1079,7 +1079,7 @@ void KarkkainenBlockwiseSA<TStr>::nextBlock(int cur_block, int tid) {
 			if(!first) {
 				// Not the first bucket
 				assert_gt(cur_block, 0);
-				assert_leq(cur_block, _sampleSuffs.size());
+				assert_leq((size_t)cur_block, _sampleSuffs.size());
 				lo = _sampleSuffs[cur_block-1];
 				//zLo.resizeExact(_dcV);
 				//zLo.fillZero();
